fixed configurations and scripts to not include editor specific code in builds
