+++ Packages/com.vrsys.core/Editor/Scripts/UserRoleSelectorDrawer.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 1248274761254c24ca58b0a4bfc4faf0
