+++ Packages/com.vrsys.core/Editor/vrsys.core.Editor.asmdef
@@ -1,4 +1,18 @@
 {
-	"name": "vrsys.core.Editor",
-	"references":[ "GUID:9d44f9d7f3932da4c84af6164f11d5e5" ]
+    "name": "vrsys.core.Editor",
+    "rootNamespace": "",
+    "references": [
+        "GUID:9d44f9d7f3932da4c84af6164f11d5e5"
+    ],
+    "includePlatforms": [
+        "Editor"
+    ],
+    "excludePlatforms": [],
+    "allowUnsafeCode": false,
+    "overrideReferences": false,
+    "precompiledReferences": [],
+    "autoReferenced": true,
+    "defineConstraints": [],
+    "versionDefines": [],
+    "noEngineReferences": false
 }
