+++ Packages/com.vrsys.meta-integrations/Editor/vrsys.meta-integrations.Editor.asmdef
@@ -4,7 +4,9 @@
     "references": [
         "GUID:ac01565a04399734bb677048f126c135"
     ],
-    "includePlatforms": [],
+    "includePlatforms": [
+        "Editor"
+    ],
     "excludePlatforms": [],
     "allowUnsafeCode": false,
     "overrideReferences": false,
